pyproject.toml
[project]
name = "uncertainty_estimation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "rubenjr0", email = "rubenjr0@proton.me" }
]
requires-python = ">=3.12"
dependencies = [
    "depth-pro",
    "lightning>=2.5.1",
    "neptune>=1.13.0",
    "rerun-sdk>=0.22.1",
    "scipy>=1.15.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]

[project.scripts]
train = "uncertainty_estimation:train"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[tool.uv.sources]
depth-pro = { git = "https://github.com/apple/ml-depth-pro" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
